test: :test_tube: added test for convertTree
